moved test class to virt/test_virt_metrics.py
